# swagger-docs.json
{"openapi":"3.0.0","info":{"title":"Marble Store Service","description":"The API to create products on the service","version":"1.0","contact":{}},"tags":[{"name":"products","description":""}],"servers":[],"components":{"schemas":{"CreateMaterialDto":{"type":"object","properties":{"name":{"type":"string","maxLength":50},"purchasePrice":{"type":"number"},"sellPrice":{"type":"number"},"inStock":{"type":"number"},"onSale":{"type":"boolean","default":false},"unitOfMeasurement":{"type":"string","default":"m²"},"shopID":{"type":"string","default":"1"}},"required":["name","purchasePrice","sellPrice","shopID"]},"UpdateMaterialDto":{"type":"object","properties":{"name":{"type":"string","maxLength":50},"purchasePrice":{"type":"number"},"sellPrice":{"type":"number"},"inStock":{"type":"number"},"onSale":{"type":"boolean","default":false},"unitOfMeasurement":{"type":"string","default":"m²"},"shopID":{"type":"string","default":"1"}},"required":["name","purchasePrice","sellPrice","shopID"]}}},"paths":{"/material":{"post":{"operationId":"MaterialController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateMaterialDto"}}}},"responses":{"201":{"description":""}},"tags":["materials"]},"get":{"operationId":"MaterialController_findAll","parameters":[{"name":"_id","required":true,"in":"query","schema":{"type":"string"}},{"name":"shopID","required":true,"in":"query","schema":{"default":"ShopID","type":"string"}},{"name":"name","required":true,"in":"query","maxLength":50,"schema":{"type":"string"}},{"name":"unitOfMeasurement","required":true,"in":"query","schema":{"default":"m²","type":"string"}},{"name":"purchasePrice","required":true,"in":"query","schema":{"type":"number"}},{"name":"sellPrice","required":true,"in":"query","schema":{"type":"number"}},{"name":"inStock","required":true,"in":"query","schema":{"type":"boolean"}},{"name":"onSale","required":true,"in":"query","schema":{"default":false,"type":"number"}}],"responses":{"200":{"description":""}},"tags":["materials"]}},"/material/{id}":{"get":{"operationId":"MaterialController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["materials"]},"put":{"operationId":"MaterialController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateMaterialDto"}}}},"responses":{"200":{"description":""}},"tags":["materials"]},"delete":{"operationId":"MaterialController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["materials"]}}}}